rm consoles
