+++ src/binding/GraphManager.ts
@@ -36,12 +36,9 @@ export class GraphManager {
         const visited = new Set<string>();
 
         const addNode = (component: BaseComponent) => {
-            console.log('adding node', component.id,nodes)
             if(nodes.find(node => node.id === component.id)) {
-                console.log('node already exists', component.id)
                 return;
             } else {
-                console.log('setting node', component.id,nodes)
                 nodes.push({
                     id: component.id,
                     type: component.constructor.name,
@@ -78,7 +75,6 @@ export class GraphManager {
 
         traverse(startComponentId);
 
-        console.log('generated binding graph', JSON.parse(JSON.stringify(nodes)), JSON.parse(JSON.stringify(edges)))
         return { nodes, edges };
     }
